create a catagory page
