+++ src/App.js
@@ -13,7 +13,8 @@ function App() {
 
       <Routes>
         <Route path="/" element={<Hero></Hero>}></Route>
-        <Route path="/new/*" element={<Page></Page>}></Route>
+        <Route path="/new" element={<Page></Page>}></Route>
+        <Route path="/catagories/*" element={<Category></Category>}></Route>
       </Routes>
     </>
   );
